+++ custom_components/evn_power_insights/const.py
@@ -2,14 +2,14 @@
 
 from datetime import timedelta
 
-DEFAULT_SCAN_INTERVAL = timedelta(hours=3)
+DEFAULT_SCAN_INTERVAL = timedelta(hours=24)
 
 DOMAIN = "evn_power_insights"
 
 CONF_DEVICE_NAME = "EVN Power Insights"
 CONF_DEVICE_MODEL = "EVN Power Insights"
 CONF_DEVICE_MANUFACTURER = "EVN Power Insights"
-CONF_DEVICE_SW_VERSION = "0.0.8"
+CONF_DEVICE_SW_VERSION = "0.0.9"
 
 CONF_USERNAME = "username"
 CONF_PASSWORD = "password"